[BUGFIX] Switched from data type SMALLINT(6) of tt_content.colPos to INT(11) to be closer to TYPO3 core and reduce issues when using it with EXT:flux See issue #36

+++ ext_tables.sql
@@ -40,7 +40,7 @@ CREATE TABLE tx_gridelements_backend_layout (
 # Table structure for table 'tt_content'
 #
 CREATE TABLE tt_content (
-	colPos smallint(6) DEFAULT '0' NOT NULL,
+	colPos int(11) DEFAULT '0' NOT NULL,
 	layout varchar(255) DEFAULT '' NOT NULL,
 	backupColPos smallint(6) DEFAULT '-2' NOT NULL,
 	tx_gridelements_backend_layout varchar(255) DEFAULT '' NOT NULL,
